Use a ThreadLocal instead a synchronization mechanism
This showed in the profile as a big time consumer. I initially removed
the whole pooling and it was 30% faster.

I tested the memory footprint and was 5 extra allocations per op:

 1 for Machine, 2 for Queue, and 2 for Queue.Entry.

So instead I used a ThreadLocal of Machine. This is the same
optimization that Teradata fork does: https://github.com/Teradata/re2j

I also removed the reset method since it is not used and doesn't make
sense after this change.