Avoid locking for put methods for RE2. Fixes #46 (#121)
* Avoid locking for put methods for RE2.

Google-wide-profiling indicated that this was a significant source of java lock contention. This new approach uses a Treiber stack to make adding an operation back into the pool a lock-free operation. It uses the existing objects as nodes in the linked stack - the Treiber stack suffers from an ABA problem when popping if nodes are reused, so removing an item from the pool is done by moving the whole stack to a simple linked stack guarded by the existing lock.

The locking could be avoided entirely by allocating wrapper objects for the stack nodes, but I'm not sure if that's desirable, since the goal of the pool was to avoid allocation.

* Address comments

* Address comments

* fix commnt

* Address comments

* Update GWT dependency to support AtomicReference

Co-authored-by: Charles Munger <clm@google.com>