fix repeat of preferred empty match
Ports fix of https://github.com/golang/go/issues/46123 to RE2/J.

Original go fix:

https://github.com/golang/go/commit/2a61b3c59088115245d084d5ae07dd4be5fbe1b0

Change description from that commit:

In Perl mode, (|a)* should match an empty string at the start of the
input. Instead it matches as many a's as possible.
Because (|a)+ is handled correctly, matching only an empty string,
this leads to the paradox that e* can match more text than e+
(for e = (|a)) and that e+ is sometimes different from ee*.

The current code treats e* and e+ as the same structure, with
different entry points. In the case of e* the preference list ends up
not quite in the right order, in part because the "before e" and
"after e" states are the same state. Splitting them apart fixes the
preference list, and that can be done by compiling e* as if it were
(e+)?.

Fixes https://github.com/google/re2j/issues/136.