Reduce the incidence of infinite loops while case folding
dc7d6e5d41225dc0825ea6fe4c6055ff854abe13 unfortunately increases the
incidence of infinite loops during case folding if re2j is running on a
JVM newer than the version used to generate the bundled
UnicodeTables.java and the input contains a rune that would require
special case folding rules to form a closed fold loop. \u1C80 (Cyrillic
Small Letter Rounded Ve) is an example of such a rune.

Workaround the issue by inverting the order of parameters passed to
equalsIgnoreCase() so that the rune from the pattern being matched,
rather than the input content, undergoes case folding instead. This does
not fully eliminate the possibility of an infinite loop in this
scenario, since the pattern may well contain one of the problematic
runes, but it effectively restores the situation as it was pre
dc7d6e5d41225dc0825ea6fe4c6055ff854abe13, since the previous logic also
performed case folding on the rune from the pattern and not on the
content.

Signed-off-by: Máté Szabó <mszabo@fandom.com>