Resolve all groups when performing match
RE2J does not resolve groups when `Matcher.find()` is called, but rater it only resolves groups when `Matcher.group(*)` is called. This results in input being matched twice reducing performance. This fix enables RE2J to resolve the groups when `Matcher.find()` is called, so that when `Matcher.group(*)` is called the results can be served from the cache.