Add `Pattern.programSize()` and `Matcher.programSize()` (#180)
* Add `Pattern.programSize()` and `Matcher.programSize()`

This PR exposes `Pattern.programSize()` and `Matcher.programSize()`
public API.

The program size represents a very approximate measure of a
regexp's "cost". Larger numbers are more expensive than smaller
numbers.

Similar to the canonical C++ implementation, re2j will return the
program size as the number of instructions of the regex program
without making any promises or claims except "larger is more
expensive".

Context: The need for this change arose from cross-language projects,
such as gRPC and CEL. gRPC needs to configure the maximum size of
regex programs in CEL to the same number across all languages. While
it's possible in CEL-Cpp and gRPC-Cpp, CEL-Java doesn't provide the
same configuration option simply because the program size is not
available in re2j. In Go, the number of instruction is available in
Go via the length of https://pkg.go.dev/regexp/syntax#Prog.Inst.

* googleJavaFormat